+++ src/lib.rs
@@ -101,6 +101,7 @@ pub use crate::algorithms::connectivity::reachability::count_reachable;
 pub use crate::algorithms::connectivity::reachability_matrix::reachability_matrix;
 pub use crate::algorithms::connectivity::strong::strongly_connected_components;
 pub use crate::algorithms::connectivity::transitive_closure::transitive_closure;
+pub use crate::algorithms::constructors::generalized_petersen::generalized_petersen;
 pub use crate::algorithms::constructors::hamming::hamming;
 pub use crate::algorithms::constructors::hypercube::{MAX_HYPERCUBE_DIMENSION, hypercube};
 pub use crate::algorithms::constructors::kary_tree::{TreeMode, kary_tree};